services/vanna/requirements.txt
vanna
fastapi
uvicorn[standard]
python-dotenv
psycopg-binary
groq